♻️ Extract component (#365)
