+++ src/routes/users/edit/+page.svelte
@@ -1,18 +1,16 @@
 <script lang="ts">
-  import { Tabs, TabItem, Input, Button, Alert, Checkbox, Li, List } from 'flowbite-svelte';
-  // @ts-ignore
-  import InfoCircleSolid from 'flowbite-svelte-icons/InfoCircleSolid.svelte';
+  import { Tabs, TabItem, Alert } from 'flowbite-svelte';
 
   import AtCoderUserValidationForm from '$lib/components/AtCoderUserValidationForm.svelte';
+  import UserAccountDeletionForm from '$lib/components/UserAccountDeletionForm.svelte';
   import ContainerWrapper from '$lib/components/ContainerWrapper.svelte';
   import FormWrapper from '$lib/components/FormWrapper.svelte';
   import ReadOnlyLabel from '$lib/components/ReadOnlyLabel.svelte';
   import SubmissionButton from '$lib/components/SubmissionButton.svelte';
 
   import { Roles } from '$lib/types/user';
-  //import type { ActionForm } from './$types';
+
   export let data;
-  //export let form;
 
   let role = data.role;
   let username = data.username;
@@ -21,8 +19,6 @@
   let atcoder_is_validated = data.is_validated;
   let message = data.message;
   let message_type = data.message_type;
-
-  let showDeleteButton = false;
 
   export let status = 'nothing';
   if (data.is_validated === true) {
@@ -34,10 +30,6 @@
 
   const isGeneralUser = (userRole: Roles, userName: string) => {
     return userRole === Roles.USER && userName !== 'guest';
-  };
-
-  const toggleDeleteButton = () => {
-    showDeleteButton = !showDeleteButton;
   };
 </script>
 
@@ -105,32 +97,7 @@
     {#if isGeneralUser(role, username)}
       <TabItem>
         <span slot="title" class="text-lg">アカウント削除</span>
-        <!-- TODO: コンポーネントとして切り出す -->
-        <ContainerWrapper>
-          <FormWrapper action="?/delete">
-            <Alert color="red" class="!items-start">
-              <InfoCircleSolid slot="icon" class="w-6 h-6" />
-              <span class="font-medium text-lg">警告</span>
-              <p class="font-medium">以下の内容が削除されます</p>
-
-              <List class="mt-1.5 ms-4 list-disc list-inside">
-                <Li>アカウント</Li>
-                <Li>登録した回答状況</Li>
-                <Li>AtCoder IDとの連携</Li>
-              </List>
-            </Alert>
-
-            <!-- hiddenでusernameを持つのは共通 -->
-            <Input size="sm" type="hidden" name="username" bind:value={username} />
-
-            <ReadOnlyLabel labelName="ユーザ名" inputValue={username} />
-            <Checkbox on:click={toggleDeleteButton}>同意する</Checkbox>
-
-            {#if showDeleteButton}
-              <Button readonly type="submit" class="w-full">アカウントを削除</Button>
-            {/if}
-          </FormWrapper>
-        </ContainerWrapper>
+        <UserAccountDeletionForm {username} />
       </TabItem>
     {/if}
   </Tabs>
